requirements_3.txt
cycler==0.10.0
greenlet==1.1.1
importlib-metadata==4.6.4
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.3.1
psutil==5.8.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.23
typing-extensions==3.10.0.0
zipp==3.5.0

